Update TalkFormatter
 - Made an interface in case we switch/use a service provider
 - Moved it to the Talk map in the Domain map
 - Updated all references accordingly
 - Moved tests out of the talk model test
 - Added test for the format list function
 - Implemented return types
 - Implemented type declarations
 - getTalkMeta now always returns a TalkMeta object
