+++ classes/Http/Controller/Admin/DashboardController.php
@@ -7,7 +7,7 @@ use OpenCFP\Domain\Model\Favorite;
 use OpenCFP\Domain\Model\Talk;
 use OpenCFP\Domain\Model\User;
 use OpenCFP\Domain\Services\Authentication;
-use OpenCFP\Domain\Services\TalkFormatter;
+use OpenCFP\Domain\Talk\TalkFormatter;
 use OpenCFP\Http\Controller\BaseController;
 
 class DashboardController extends BaseController
